split up search components
