+++ packages/lesmis/src/utils/index.js
@@ -5,7 +5,8 @@
  */
 export * from './makeEntityLink';
 export * from './search';
-export * from './useUrlParams';
-export * from './useUrlSearchParams';
+export * from './useAutocomplete';
 export * from './useEntityBreadcrumbs';
 export * from './useExportToPNG';
+export * from './useUrlParams';
+export * from './useUrlSearchParams';
